小说预渲染 MIP 核心支持 (#320)
