+++ docs/guide/all-sites-mip/meta.json
@@ -16,6 +16,9 @@
       "key": "mip-shell"
     },
     {
+      "key": "performace"
+    },
+    {
       "key": "viewer"
     },
     {
