+++ docs/guide/all-sites-mip/performance.md
@@ -1,0 +1,76 @@
+## 提升性能
+
+MIP 的本意是为了提升 WEB 站点访问的体验。在页面切换时通过增加切换动画，避免白屏，给使用者带来了更好的体验，获得了“感官上的性能提升”。我们是否有办法在实质上提升性能，而非仅仅是“感官”和“体验”呢？
+
+### 复用已有页面
+
+在打开新的 MIP 页面时（注意：不包括后退），为了保证页面是最新的，即使以前曾经打开过这个页面，也依然会删除旧的页面，而重新打开并载入这个页面。
+
+最熟悉站点的人应该是站长以及站点本身的开发人员。如果一个站点的内容并不经常修改，MIP 允许站点通过牺牲时效性换取更好的性能。例如小说阅读站点，新闻站点等均在此列。
+
+实现原理也很简单，在打开新页面时，如果这个页面已经在之前加载过（即容器 iframe 已经存在，只是隐藏了而已），则不重新和创建和载入，而是直接使用这个页面。和普通的打开新页面相比，复用已有页面能够获得三个优势：
+
+1. 不发送网络请求，因此没有网络延时，__速度更快__
+
+2. 切换时，因为目标页面已经存在，所以 __没有 Loading__ 图标，体验更加顺畅
+
+3. 因为只是直接把目标页面展现出来，所以 __能够保留目标页面的状态和滚动位置__
+
+使用方法非常简单，只需要在切换页面的入口 (`<a>` 链接) 增加 `cache-first` 的属性，如下：
+
+```html
+<a mip-link cache-first href="http://localhost:8080/examples/page/tree.html">Go to Tree</a>
+```
+
+或者使用编码形式：
+
+```javascript
+window.MIP.viewer.open('http://localhost:8080/examples/page/tree.html', {
+    cacheFirst: true
+})
+```
+
+如果目标页面之前没有加载过（或者限于 iframe 的最大数量已被清除），MIP 会忽略这个参数，进入普通的打开页面的流程，因此这个功能没有副作用，可以放心使用。
+
+### 预渲染
+
+和普通切换相比，在复用已有页面时获得的性能提升非常大，但需要牺牲一定的时效性，因为开发者并不清楚用户上一次打开目标页面是在什么时候；另外“必须是已经打开过的页面”也是一个有些严苛的要求。
+
+因此 MIP 另外推出一种机制称为“预渲染”。它主要分为两个步骤：
+
+1. 给开发者提供一个方法，获取需要预渲染的页面 URL 列表，逐个创建它们的容器 iframe，但保留隐藏状态，__且阻断所有组件的生命周期__
+
+2. 使用和“复用已有页面”相同的 `cache-first` 属性，从已有的（预渲染的）页面直接调出显示
+
+之所以要阻断目标页面的生命周期，是因为有些组件需要请求网络资源，或者占用计算资源等。毕竟它们是属于预渲染的页面，因此存在“预渲染了但最终用户没有打开”的可能性，这样这些资源就被浪费了，就可能造成无谓的网络流量消耗，页面卡顿，电池浪费等等。__如果有些组件必须不能被阻断，要求在预渲染情况下也要正常执行的，可以添加 `prerender` 属性，详见[这里](https://github.com/mipengine/mip2/issues/284#issuecomment-423026357)。__
+
+在使用体验方面，预渲染能够获得和复用已有页面相同的优势，即速度快和没有 Loading。（因为预渲染新页面，所以滚动位置总在最上方）
+
+开发者可以在当前页面加载完成后（或者其他任何他们认为合适的时机）通过 `window.MIP.viewer.page.prerender` 方法加载预渲染页面（这个方法不要求必须在 Root Page 调用），这个方法参数如下：
+
+* urls: __Array|string__, 需要预渲染的页面 URL，__必须是包含协议，域名，端口号等的完整 URL 而不是相对或者绝对路径，不支持跨域__。如果是单个页面，可以是字符串，否则是字符串组成的数组。
+
+示例代码：
+
+```javascript
+window.MIP.viewer.page.prerender([
+    'http://localhost:8080/examples/page/tree.html',
+    'http://localhost:8080/examples/page/data.html'
+])
+```
+
+在预渲染之后，和复用已有页面相同，开发者可以通过在切换页面的入口 (`<a>` 链接) 增加 `cache-first` 的属性，如下：
+
+```html
+<a mip-link cache-first href="http://localhost:8080/examples/page/tree.html">Go to Tree</a>
+```
+
+或者使用编码形式：
+
+```javascript
+window.MIP.viewer.open('http://localhost:8080/examples/page/tree.html', {
+    cacheFirst: true
+})
+```
+
+注意：为了页面性能和内存大小考虑，MIP 允许页面上存在的 iframe 个数是有上限的（目前是 6 个）。在加载新的 iframe 触及上限时，MIP 会删除页面上已有的 iframe（当然会避开当前页面和目标页面）。这里会优先清除普通页面，保留预缓存页面。而如果已经没有普通页面可以删除时，则预缓存页面也可能被删除。因此如果一下子预缓存太多的页面，可能反而无法命中。
