+++ packages/mip/src/page/index.js
@@ -4,6 +4,7 @@
  */
 
 import {
+  createIFrame,
   ensureMIPShell,
   getIFrame,
   toggleFadeHeader
@@ -21,9 +22,10 @@ import {
   MESSAGE_ROUTER_BACK,
   MESSAGE_ROUTER_FORWARD,
   MESSAGE_CROSS_ORIGIN,
-  MESSAGE_BROADCAST_EVENT
+  MESSAGE_BROADCAST_EVENT,
+  DEFAULT_SHELL_CONFIG
 } from './const/index'
-
+import fn from '../util/fn'
 import {customEmit} from '../util/custom-event'
 import viewport from '../viewport'
 import performance from '../performance'
@@ -41,11 +43,11 @@ class Page {
     this.pageId = undefined
     this.fullpath = undefined
     this.pageMeta = undefined
-
-    // root page
-    this.children = []
     this.currentPageId = undefined
     this.targetWindow = window
+
+    // 记录 iframe 内的 page 对象。Root Page 这个属性才有意义，但 Root Page 本身不计入。
+    this.children = []
   }
 
   initPageId () {
@@ -297,6 +299,25 @@ class Page {
     })
   }
 
+  /**
+   * Create and prerender iframe(s) sliently
+   * Cross Origin is now allowed
+   * Each page can invoke this function
+   *
+   * @param {Array|string} urls
+   * @returns {Promise}
+   */
+  prerender (urls) {
+    /* istanbul ignore next */
+    if (this.isCrossOrigin) {
+      console.warn('跨域 MIP 页面暂不支持预渲染')
+      return
+    }
+
+    let target = this.isRootPage ? this : /* istanbul ignore next */ window.parent.MIP.viewer.page
+    return target.prerenderPages(urls)
+  }
+
   // =============================== Root Page methods ===============================
 
   /**
@@ -305,6 +326,11 @@ class Page {
    * @param {Object} event event
    */
   emitEventInCurrentPage (event) {
+    /* istanbul ignore next */
+    if (!this.isRootPage) {
+      console.warn('该方法只能在 rootPage 调用')
+      return
+    }
     let currentPage = this.getPageById(this.currentPageId)
     this.emitCustomEvent(currentPage.targetWindow, currentPage.isCrossOrigin, event)
   }
@@ -315,6 +341,11 @@ class Page {
    * @param {Page} page page
    */
   addChild (page) {
+    /* istanbul ignore next */
+    if (!this.isRootPage) {
+      console.warn('该方法只能在 rootPage 调用')
+      return
+    }
     for (let i = 0; i < this.children.length; i++) {
       if (this.children[i].pageId === page.pageId) {
         this.children.splice(i, 1)
@@ -325,25 +356,52 @@ class Page {
   }
 
   /**
-   * check if children.length exceeds MAX_PAGE_NUM
-   * if so, remove the first child
+   * Check if children.length exceeds MAX_PAGE_NUM
+   * If so, remove one
+   *
+   * @param {string} targetPageId targetPageId
    */
   /* istanbul ignore next */
   checkIfExceedsMaxPageNum (targetPageId) {
+    /* istanbul ignore next */
+    if (!this.isRootPage) {
+      console.warn('该方法只能在 rootPage 调用')
+      return
+    }
     if (this.children.length >= MAX_PAGE_NUM) {
       let currentPage
+      let prerenderIFrames = []
+      let found = false
       for (let i = 0; i < this.children.length; i++) {
         currentPage = this.children[i]
-        // find first removable page, which can't be target page or current page
+        // Find first removable page, which can't be target page or current page
         if (currentPage.pageId !== targetPageId &&
           currentPage.pageId !== this.currentPageId) {
           const firstRemovableIframe = getIFrame(currentPage.pageId)
+
+          // If prerendered, skip it first
+          if (firstRemovableIframe.getAttribute('prerender') === '1') {
+            prerenderIFrames.push({iframe: firstRemovableIframe, index: i})
+            continue
+          }
+
           if (firstRemovableIframe && firstRemovableIframe.parentNode) {
             firstRemovableIframe.parentNode.removeChild(firstRemovableIframe)
+            this.children.splice(i, 1)
+            found = true
           }
-          // remove from children list
-          this.children.splice(i, 1)
           return
+        }
+      }
+
+      if (!found) {
+        // Find one in prerendered iframe
+        for (let i = 0; i < prerenderIFrames.length; i++) {
+          const firstRemovableIframe = prerenderIFrames[i].iframe
+          if (firstRemovableIframe && firstRemovableIframe.parentNode) {
+            firstRemovableIframe.parentNode.removeChild(firstRemovableIframe)
+            this.children.splice(prerenderIFrames[i].index, 1)
+          }
         }
       }
     }
@@ -356,6 +414,11 @@ class Page {
    * @return {Page} page
    */
   getPageById (pageId) {
+    /* istanbul ignore next */
+    if (!this.isRootPage) {
+      console.warn('该方法只能在 rootPage 调用')
+      return
+    }
     if (!pageId || pageId === this.pageId) {
       return this
     }
@@ -375,6 +438,11 @@ class Page {
    * @return {Array<HTMLElement>} elements
    */
   getElementsInRootPage () {
+    /* istanbul ignore next */
+    if (!this.isRootPage) {
+      console.warn('该方法只能在 rootPage 调用')
+      return
+    }
     let whitelist = [
       '.mip-page__iframe',
       '.mip-page-loading-wrapper',
@@ -390,6 +458,104 @@ class Page {
     ]
     let notInWhitelistSelector = whitelist.map(selector => `:not(${selector})`).join('')
     return [...document.querySelectorAll(`body > ${notInWhitelistSelector}`)]
+  }
+
+  /**
+   * Create and prerender iframe(s) sliently
+   * Cross Origin is now allowed
+   *
+   * @param {Array|string} urls
+   * @returns {Promise}
+   */
+  prerenderPages (urls) {
+    /* istanbul ignore next */
+    if (!this.isRootPage) {
+      console.warn('该方法只能在 rootPage 调用')
+      return Promise.reject()
+    }
+    if (typeof urls === 'string') {
+      urls = [urls]
+    }
+
+    if (!Array.isArray(urls)) {
+      return Promise.reject('预渲染参数必须是一个数组')
+    }
+
+    let createPrerenderIFrame = ({fullpath, pageId}) => {
+      return new Promise((resolve, reject) => {
+        let me = this
+        let iframe = getIFrame(pageId)
+        /* istanbul ignore next */
+        if (iframe) {
+          // 预加载前已经存在，直接返回即可
+          resolve(iframe)
+          return
+        }
+
+        createIFrame({fullpath: fullpath + '#prerender=1', pageId}, {
+          onLoad (newIframe) {
+            newIframe.setAttribute('prerender', '1')
+            let targetPageInfo = {
+              pageId,
+              pageMeta: fn.extend(true, {}, findMetaByPageId(pageId)),
+              fullpath,
+              standalone: window.MIP.standalone,
+              isRootPage: false,
+              isCrossOrigin: false,
+              isPrerender: true
+            }
+            targetPageInfo.targetWindow = newIframe.contentWindow
+            me.addChild(targetPageInfo)
+            me.checkIfExceedsMaxPageNum(pageId)
+
+            resolve(newIframe)
+          },
+          onError (newIframe) {
+            /* istanbul ignore next */
+            reject(newIframe)
+          }
+        })
+      })
+    }
+
+    let findMetaByPageId = pageId => {
+      let target
+      /* istanbul ignore next */
+      if (!this.isRootPage && !this.isCrossOrigin) {
+        target = window.parent
+      } else {
+        target = window
+      }
+
+      /* istanbul ignore next */
+      if (target.MIP_PAGE_META_CACHE[pageId]) {
+        return target.MIP_PAGE_META_CACHE[pageId]
+      } else {
+        for (let i = 0; i < target.MIP_SHELL_CONFIG.length; i++) {
+          let route = target.MIP_SHELL_CONFIG[i]
+          if (route.regexp.test(pageId)) {
+            target.MIP_PAGE_META_CACHE[pageId] = route.meta
+            return route.meta
+          }
+        }
+      }
+
+      /* istanbul ignore next */
+      console.warn('Cannot find MIP Shell Config for current page. Use default instead.')
+      /* istanbul ignore next */
+      return Object.assign({}, DEFAULT_SHELL_CONFIG)
+    }
+
+    return Promise.all(urls.map(fullpath => {
+      /* istanbul ignore next */
+      if (window.MIP.viewer._isCrossOrigin(fullpath)) {
+        console.warn('跨域 MIP 页面暂不支持预渲染', fullpath)
+        return Promise.resolve()
+      }
+
+      let pageId = getCleanPageId(fullpath)
+      return createPrerenderIFrame({fullpath, pageId})
+    }))
   }
 }
