+++ packages/mip/src/styles/mip-page.less
@@ -1,4 +1,5 @@
 @import "./variable.less";
+
 body::before {
   display: block;
   content: '';
